==> notebooks/experiments/exp_01/01_preprocessing_A.ipynb <==
import pandas as pd
import pickle as pkl
from numpy import savetxt
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import LabelEncoder

# Params
target = 'Survived'
features = "Sex"
idx = 'Passengerid'

# Paths 
label_enconder_path = "/home/walter/Documents/personal_projects/new-titan/notebooks/experiments/exp_01/checkpoint/label_encoder.pkl"

def train_label_encoder(df):
    encoder = LabelEncoder()
    encoder.fit(df)

    enc_output = open(label_enconder_path, "wb")
    pkl.dump(encoder, enc_output)
    enc_output.close()

    print("Encoder saved in:{}".format(label_enconder_path))

def transform_label_encoder(df):
    pkl_file = open(label_enconder_path, 'rb')
    encoder = pkl.load(pkl_file) 
    pkl_file.close()

    res = encoder.transform(df)

    return res

def save_train_files():
    savetxt('/home/walter/Documents/personal_projects/new-titan/data/processed/data_train/X_train.csv', X_train, delimiter=',')
    savetxt('/home/walter/Documents/personal_projects/new-titan/data/processed/data_train/y_train.csv', y_train, delimiter=',')
    savetxt('/home/walter/Documents/personal_projects/new-titan/data/processed/data_train/label_train.csv', label_train, delimiter=',')
    savetxt('/home/walter/Documents/personal_projects/new-titan/data/processed/data_test/label_test.csv', label_test, delimiter=',')
    savetxt('/home/walter/Documents/personal_projects/new-titan/data/processed/data_test/X_test.csv', X_test, delimiter=',')
    savetxt('/home/walter/Documents/personal_projects/new-titan/data/processed/data_test/y_test.csv', y_test, delimiter=',')

# load
data = pd.read_csv("/home/walter/Documents/personal_projects/new-titan/data/raw/train.csv", header=0)

# Preprocessing
# 1. Divide train-test
X_train, X_test, y_train, y_test = train_test_split(data[features], data[target], test_size=0.1, random_state=100)
label_train = data.iloc[X_train.index, 0]
label_test = data.iloc[X_test.index, 0]

# 2. encoding
train_label_encoder(X_train)
X_train = transform_label_encoder(X_train)
X_test = transform_label_encoder(X_test)

# Save
save_train_files()
