==> notebooks/experiments/exp_01/02_model_preview.ipynb <==
import numpy as np
import os
import pandas as pd
import pickle
from sklearn.tree import DecisionTreeClassifier
from sklearn.model_selection import KFold
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score

# Paths
prefix = '/home/walter/Documents/personal_projects/new-titan/data/processed'
selected_model_path = '/home/walter/Documents/personal_projects/new-titan/notebooks/experiments/exp_01/artifacts/selected_model.pkl'

# defino función load
def load_data(prefix):
    X_train = np.genfromtxt(os.path.join(prefix, 'data_train', 'X_train.csv'), delimiter=',')
    y_train = np.genfromtxt(os.path.join(prefix, 'data_train', 'y_train.csv'), delimiter=',').astype('int')
    label_train = np.genfromtxt(os.path.join(prefix, 'data_train', 'label_train.csv'), delimiter=',')
    X_test = np.genfromtxt(os.path.join(prefix, 'data_test', 'X_test.csv'), delimiter=',')
    y_test = np.genfromtxt(os.path.join(prefix, 'data_test', 'y_test.csv'), delimiter=',').astype('int')
    label_test = np.genfromtxt(os.path.join(prefix, 'data_test', 'label_test.csv'), delimiter=',')
   
    return X_train, y_train, label_train, X_test, y_test, label_test

# defino función de eval
def eval_model(X, y, model):
    k = 5
    kf = KFold(n_splits=k, shuffle=True, random_state=2)

    acc_score = []

    for train_index, valid_index in kf.split(X):

        # load
        X_train, X_valid = X[train_index], X[valid_index]
        y_train, y_valid = y[train_index], y[valid_index]

        # model
        model.fit(X_train.reshape(-1,1), y_train)
        y_predict = model.predict(X_valid.reshape(-1,1))

        # eval
        acc = accuracy_score(y_predict, y_valid)
        acc_score.append(acc)

    avg_acc_score = sum(acc_score)/k
    return avg_acc_score    

# load

# El ejercicio es el siguiente: se utiliza el 90% denominado train para realizar
# el cv. y seleccion de modelo. El 10% se utilizará para evaluación del modelo.

# No se considera el siguiente problema: el preprocesamiento 
# 01_preprocessing debió realizarse DENTRO DEL CV.
# Aquí no habrá problemas pero queda para próximas iteraciones.

X, y, label, _, _, _ = load_data(prefix)

# eval
models = [
    DecisionTreeClassifier(random_state = 2),
    LogisticRegression(random_state=2)
]

for model in models:
    print(eval_model(X, y, model))

# Model selection
selected_model = models[0]

with open(selected_model_path, 'wb') as out:
    pickle.dump(model, out)

model
