==> notebooks/experiments/exp_01/03_tunning.ipynb <==
import os
import numpy as np
import pandas as pd
import pickle
from sklearn.model_selection import train_test_split
from sklearn.model_selection import RandomizedSearchCV

# Paths
prefix = '/home/walter/Documents/personal_projects/new-titan/data/processed'
selected_model_path = '/home/walter/Documents/personal_projects/new-titan/notebooks/experiments/exp_01/artifacts/selected_model.pkl'
tunned_model_path = '/home/walter/Documents/personal_projects/new-titan/notebooks/experiments/exp_01/artifacts/tunned_model.pkl'

# Tunning
solver = ['newton-cg', 'lbfgs', 'liblinear', 'sag', 'saga']
C = [100, 10, 1, 0.1, 0.001]
cv = 5
n_iter = 20
score = 'accuracy'

def load_model(path):
    pkl_file = open(path, 'rb')
    selected_model = pickle.load(pkl_file)
    pkl_file.close()

    return selected_model

def load_data(prefix):
    X_train = np.genfromtxt(os.path.join(prefix, 'data_train', 'X_train.csv'), delimiter=',')
    y_train = np.genfromtxt(os.path.join(prefix, 'data_train', 'y_train.csv'), delimiter=',').astype('int')
    label_train = np.genfromtxt(os.path.join(prefix, 'data_train', 'label_train.csv'), delimiter=',')
    X_test = np.genfromtxt(os.path.join(prefix, 'data_test', 'X_test.csv'), delimiter=',')
    y_test = np.genfromtxt(os.path.join(prefix, 'data_test', 'y_test.csv'), delimiter=',').astype('int')
    label_test = np.genfromtxt(os.path.join(prefix, 'data_test', 'label_test.csv'), delimiter=',')
   
    return X_train, y_train, label_train, X_test, y_test, label_test

# load model and data
model = load_model(selected_model_path)
X, y, label, X_test, y_test, label_test = load_data(prefix)
model

# tunning hyper
from sklearn.model_selection import RandomizedSearchCV


params = {
    'solver': solver,
    'C':C
}

search = RandomizedSearchCV(
    estimator=model,
    param_distributions=params,
    scoring=score,
    cv=cv,
    n_iter=n_iter,
    return_train_score=False,
    random_state=100,
    refit=True
)

search.fit(X.reshape(-1,1), y)

pd.DataFrame(search.cv_results_)

with open(tunned_model_path, 'wb') as out:
    pickle.dump(search.best_estimator_, out)

search.best_estimator_
