==> notebooks/experiments/exp_01/05_train_for_prod.ipynb <==
import pandas as pd
import pickle
from numpy import savetxt
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import LabelEncoder

# Params
target = 'Survived'
features = "Sex"
idx = 'Passengerid'

# Paths 
label_enconder_path = "/home/walter/Documents/personal_projects/new-titan/notebooks/experiments/exp_01/checkpoint/label_encoder_for_prod.pkl"
tunned_model_path = '/home/walter/Documents/personal_projects/new-titan/notebooks/experiments/exp_01/artifacts/tunned_model.pkl'
trained_model_path = '/home/walter/Documents/personal_projects/new-titan/notebooks/experiments/exp_01/artifacts/model.pkl'

def train_label_encoder(df):
    encoder = LabelEncoder()
    encoder.fit(df)

    enc_output = open(label_enconder_path, "wb")
    pickle.dump(encoder, enc_output)
    enc_output.close()

    print("Encoder saved in:{}".format(label_enconder_path))

def transform_label_encoder(df):
    pkl_file = open(label_enconder_path, 'rb')
    encoder = pickle.load(pkl_file) 
    pkl_file.close()

    res = encoder.transform(df)

    return res

def load_model(path):
    pkl_file = open(path, 'rb')
    selected_model = pickle.load(pkl_file)
    pkl_file.close()

    return selected_model

# load
data = pd.read_csv("/home/walter/Documents/personal_projects/new-titan/data/raw/train.csv", header=0)

# Preprocessing
X_train = data[features]
y_train = data[target]

# encoding
train_label_encoder(X_train)
X_train = transform_label_encoder(X_train)

# fit
model = load_model(tunned_model_path)
model.fit(X_train.reshape(-1,1), y_train)

# save model
with open(trained_model_path, 'wb') as out:
    pickle.dump(model, out)

model
